refactor Sidecar::SidecarServer to separate file
