+++ sidecar/src/service/sidecar_server.rs
@@ -1,0 +1,647 @@
+use crate::config::get_product_endpoint;
+use crate::interface::{
+    AppOrQueue, RuntimeInfo, SessionConfig, SidecarAction, SidecarStats, TraceFlusher,
+};
+use crate::log;
+use crate::log::{MULTI_LOG_FILTER, MULTI_LOG_WRITER};
+use crate::service::{
+    InstanceId, QueueId, RequestIdentification, RequestIdentifier, RuntimeMetadata,
+    SerializedTracerHeaderTags, SessionInfo, SidecarInterface,
+};
+use datadog_ipc::platform::{AsyncChannel, ShmHandle};
+use datadog_ipc::tarpc;
+use datadog_ipc::tarpc::context::Context;
+use datadog_ipc::transport::Transport;
+use datadog_trace_protobuf::pb;
+use datadog_trace_utils::trace_utils;
+use datadog_trace_utils::trace_utils::SendData;
+use ddcommon::Endpoint;
+use ddtelemetry::worker::{LifecycleAction, TelemetryActions, TelemetryWorkerBuilder};
+use futures::future;
+use futures::future::{join_all, Ready};
+use manual_future::{ManualFuture, ManualFutureCompleter};
+use std::collections::hash_map::Entry;
+use std::collections::{HashMap, HashSet};
+use std::pin::Pin;
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::{Arc, Mutex, MutexGuard};
+use tracing::{debug, enabled, error, info, warn, Level};
+
+use futures::FutureExt;
+
+use datadog_ipc::platform::FileBackedHandle;
+use datadog_ipc::tarpc::server::Channel;
+
+type NoResponse = Ready<()>;
+
+fn no_response() -> NoResponse {
+    future::ready(())
+}
+
+#[derive(Default, Clone)]
+pub struct SidecarServer {
+    pub trace_flusher: Arc<TraceFlusher>,
+    sessions: Arc<Mutex<HashMap<String, SessionInfo>>>,
+    session_counter: Arc<Mutex<HashMap<String, u32>>>,
+    pub self_telemetry_config:
+        Arc<Mutex<Option<ManualFutureCompleter<ddtelemetry::config::Config>>>>,
+    pub submitted_payloads: Arc<AtomicU64>,
+}
+
+impl SidecarServer {
+    pub async fn accept_connection(self, async_channel: AsyncChannel) {
+        let server = tarpc::server::BaseChannel::new(
+            tarpc::server::Config {
+                pending_response_buffer: 10000,
+            },
+            Transport::from(async_channel),
+        );
+
+        let mut executor = datadog_ipc::sequential::execute_sequential(
+            server.requests(),
+            self.clone().serve(),
+            100,
+        );
+        let (tx, mut rx) = tokio::sync::mpsc::channel::<_>(100);
+        let tx = executor.swap_sender(tx);
+
+        let session_counter = self.session_counter.clone();
+        let submitted_payloads = self.submitted_payloads.clone();
+        let session_interceptor = tokio::spawn(async move {
+            let mut sessions = HashSet::new();
+            let mut instances = HashSet::new();
+            loop {
+                let (serve, req) = match rx.recv().await {
+                    None => return (sessions, instances),
+                    Some(s) => s,
+                };
+
+                submitted_payloads.fetch_add(1, Ordering::Relaxed);
+
+                let instance: RequestIdentifier = req.get().extract_identifier();
+                if tx.send((serve, req)).await.is_ok() {
+                    if let RequestIdentifier::InstanceId(ref instance_id) = instance {
+                        instances.insert(instance_id.clone());
+                    }
+                    if let RequestIdentifier::SessionId(session)
+                    | RequestIdentifier::InstanceId(InstanceId {
+                        session_id: session,
+                        ..
+                    }) = instance
+                    {
+                        if sessions.insert(session.clone()) {
+                            match session_counter.lock().unwrap().entry(session) {
+                                Entry::Occupied(mut entry) => entry.insert(entry.get() + 1),
+                                Entry::Vacant(entry) => *entry.insert(1),
+                            };
+                        }
+                    }
+                }
+            }
+        });
+
+        if let Err(e) = executor.await {
+            warn!("Error from executor: {e:?}");
+        }
+
+        if let Ok((sessions, instances)) = session_interceptor.await {
+            for session in sessions {
+                let stop = {
+                    let mut counter = self.session_counter.lock().unwrap();
+                    if let Entry::Occupied(mut entry) = counter.entry(session.clone()) {
+                        if entry.insert(entry.get() - 1) == 1 {
+                            entry.remove();
+                            true
+                        } else {
+                            false
+                        }
+                    } else {
+                        false
+                    }
+                };
+                if stop {
+                    self.stop_session(&session).await;
+                }
+            }
+            for instance_id in instances {
+                let maybe_session = self.lock_sessions().get(&instance_id.session_id).cloned();
+                if let Some(session) = maybe_session {
+                    session.shutdown_runtime(&instance_id.runtime_id).await;
+                }
+            }
+        }
+    }
+
+    pub fn active_session_count(&self) -> usize {
+        self.session_counter.lock().unwrap().len()
+    }
+
+    fn get_session(&self, session_id: &String) -> SessionInfo {
+        let mut sessions = self.lock_sessions();
+        match sessions.get(session_id) {
+            Some(session) => session.clone(),
+            None => {
+                let mut session = SessionInfo::default();
+                #[cfg(feature = "tracing")]
+                if enabled!(Level::INFO) {
+                    session.session_id.clone_from(session_id);
+                    info!("Initializing new session: {}", session_id);
+                }
+                sessions.insert(session_id.clone(), session.clone());
+                session
+            }
+        }
+    }
+
+    fn get_runtime(&self, instance_id: &InstanceId) -> RuntimeInfo {
+        let session = self.get_session(&instance_id.session_id);
+        session.get_runtime(&instance_id.runtime_id)
+    }
+
+    async fn stop_session(&self, session_id: &String) {
+        let session = match self.lock_sessions().remove(session_id) {
+            Some(session) => session,
+            None => return,
+        };
+
+        info!("Shutting down session: {}", session_id);
+        session.shutdown().await;
+        debug!("Successfully shut down session: {}", session_id);
+    }
+
+    fn lock_sessions(&self) -> MutexGuard<HashMap<String, SessionInfo>> {
+        self.sessions.lock().unwrap()
+    }
+
+    async fn get_app(
+        &self,
+        instance_id: &InstanceId,
+        runtime_meta: &RuntimeMetadata,
+        service_name: &str,
+        env_name: &str,
+        initial_actions: Vec<TelemetryActions>,
+    ) -> Option<crate::interface::AppInstance> {
+        let rt_info = self.get_runtime(instance_id);
+
+        let (app_future, completer) = rt_info.get_app(service_name, env_name);
+        if completer.is_none() {
+            return app_future.await;
+        }
+
+        let mut builder = TelemetryWorkerBuilder::new_fetch_host(
+            service_name.to_owned(),
+            runtime_meta.language_name.clone(),
+            runtime_meta.language_version.clone(),
+            runtime_meta.tracer_version.clone(),
+        );
+        builder.runtime_id = Some(instance_id.runtime_id.clone());
+        builder.application.env = Some(env_name.to_owned());
+        let session_info = self.get_session(&instance_id.session_id);
+        let mut config = session_info
+            .session_config
+            .lock()
+            .unwrap()
+            .clone()
+            .unwrap_or_else(ddtelemetry::config::Config::from_env);
+        config.restartable = true;
+
+        // TODO: log errors
+        let instance_option = match builder.spawn_with_config(config.clone()).await {
+            Ok((handle, worker_join)) => {
+                info!("spawning telemetry worker {config:?}");
+
+                let instance = crate::interface::AppInstance {
+                    telemetry: handle,
+                    telemetry_worker_shutdown: worker_join.map(Result::ok).boxed().shared(),
+                    telemetry_metrics: Default::default(),
+                };
+
+                instance.telemetry.send_msgs(initial_actions).await.ok();
+
+                instance
+                    .telemetry
+                    .send_msg(TelemetryActions::Lifecycle(LifecycleAction::Start))
+                    .await
+                    .ok();
+                Some(instance)
+            }
+            Err(e) => {
+                error!("could not spawn telemetry worker {:?}", e);
+                None
+            }
+        };
+        completer.unwrap().complete(instance_option).await;
+        app_future.await
+    }
+
+    fn send_trace_v04(&self, headers: &SerializedTracerHeaderTags, data: &[u8], target: &Endpoint) {
+        let headers = match headers.try_into() {
+            Ok(headers) => headers,
+            Err(e) => {
+                error!("Failed to convert SerializedTracerHeaderTags into TracerHeaderTags with error {:?}", e);
+                return;
+            }
+        };
+
+        let size = data.len();
+        let traces: Vec<Vec<pb::Span>> = match rmp_serde::from_slice(data) {
+            Ok(res) => res,
+            Err(err) => {
+                error!("Error deserializing trace from request body: {err}");
+                return;
+            }
+        };
+
+        if traces.is_empty() {
+            error!("No traces deserialized from the request body.");
+            return;
+        }
+
+        let payload =
+            trace_utils::collect_trace_chunks(traces, &headers, |_chunk, _root_span_index| {});
+
+        // send trace payload to our trace flusher
+        let data = SendData::new(size, payload, headers, target);
+        self.trace_flusher.enqueue(data);
+    }
+
+    pub async fn compute_stats(&self) -> SidecarStats {
+        let mut telemetry_stats_errors = 0;
+        let telemetry_stats = join_all({
+            let sessions = self.lock_sessions();
+            let mut futures = vec![];
+            for (_, s) in sessions.iter() {
+                let runtimes = s.lock_runtimes();
+                for (_, r) in runtimes.iter() {
+                    let apps = r.lock_apps();
+                    for (_, a) in apps.iter() {
+                        if let Some(Some(existing_app)) = a.peek() {
+                            match existing_app.telemetry.stats() {
+                                Ok(future) => futures.push(future),
+                                Err(_) => telemetry_stats_errors += 1,
+                            }
+                        }
+                    }
+                }
+            }
+            futures
+        })
+        .await;
+        let sessions = self.lock_sessions();
+        SidecarStats {
+            trace_flusher: self.trace_flusher.stats(),
+            sessions: sessions.len() as u32,
+            session_counter_size: self.session_counter.lock().unwrap().len() as u32,
+            runtimes: sessions
+                .values()
+                .map(|s| s.lock_runtimes().len() as u32)
+                .sum(),
+            apps: sessions
+                .values()
+                .map(|s| {
+                    s.lock_runtimes()
+                        .values()
+                        .map(|r| r.lock_apps().len() as u32)
+                        .sum::<u32>()
+                })
+                .sum(),
+            active_apps: sessions
+                .values()
+                .map(|s| {
+                    s.lock_runtimes()
+                        .values()
+                        .map(|r| r.lock_app_or_actions().len() as u32)
+                        .sum::<u32>()
+                })
+                .sum(),
+            enqueued_apps: sessions
+                .values()
+                .map(|s| {
+                    s.lock_runtimes()
+                        .values()
+                        .map(|r| {
+                            r.lock_app_or_actions()
+                                .values()
+                                .filter(|a| matches!(a, AppOrQueue::Queue(_)))
+                                .count() as u32
+                        })
+                        .sum::<u32>()
+                })
+                .sum(),
+            enqueued_telemetry_data: sessions
+                .values()
+                .map(|s| {
+                    s.lock_runtimes()
+                        .values()
+                        .map(|r| {
+                            r.lock_app_or_actions()
+                                .values()
+                                .filter_map(|a| match a {
+                                    AppOrQueue::Queue(q) => Some(q.stats()),
+                                    _ => None,
+                                })
+                                .sum()
+                        })
+                        .sum()
+                })
+                .sum(),
+            telemetry_metrics_contexts: sessions
+                .values()
+                .map(|s| {
+                    s.lock_runtimes()
+                        .values()
+                        .map(|r| {
+                            r.lock_apps()
+                                .values()
+                                .map(|a| {
+                                    a.peek().unwrap_or(&None).as_ref().map_or(0, |w| {
+                                        w.telemetry_metrics.lock().unwrap().len() as u32
+                                    })
+                                })
+                                .sum::<u32>()
+                        })
+                        .sum::<u32>()
+                })
+                .sum(),
+            telemetry_worker_errors: telemetry_stats_errors
+                + telemetry_stats.iter().filter(|v| v.is_err()).count() as u32,
+            telemetry_worker: telemetry_stats.into_iter().filter_map(|v| v.ok()).sum(),
+            log_filter: MULTI_LOG_FILTER.stats(),
+            log_writer: MULTI_LOG_WRITER.stats(),
+        }
+    }
+}
+
+impl SidecarInterface for SidecarServer {
+    type EnqueueActionsFut = NoResponse;
+
+    fn enqueue_actions(
+        self,
+        _context: Context,
+        instance_id: InstanceId,
+        queue_id: QueueId,
+        actions: Vec<SidecarAction>,
+    ) -> Self::EnqueueActionsFut {
+        let rt_info = self.get_runtime(&instance_id);
+        let mut queue = rt_info.lock_app_or_actions();
+        match queue.entry(queue_id) {
+            Entry::Occupied(mut entry) => match entry.get_mut() {
+                AppOrQueue::Queue(ref mut data) => {
+                    data.process(actions);
+                }
+                AppOrQueue::App(service_future) => {
+                    let service_future = service_future.clone();
+                    // drop on stop
+                    if actions.iter().any(|action| {
+                        matches!(
+                            action,
+                            SidecarAction::Telemetry(TelemetryActions::Lifecycle(
+                                LifecycleAction::Stop
+                            ))
+                        )
+                    }) {
+                        entry.remove();
+                    }
+                    let apps = rt_info.apps.clone();
+                    tokio::spawn(async move {
+                        let service = service_future.await;
+                        let app_future = if let Some(fut) = apps.lock().unwrap().get(&service) {
+                            fut.clone()
+                        } else {
+                            return;
+                        };
+                        if let Some(mut app) = app_future.await {
+                            let actions =
+                                crate::interface::EnqueuedTelemetryData::process_immediately(
+                                    actions, &mut app,
+                                )
+                                .await;
+                            app.telemetry.send_msgs(actions).await.ok();
+                        }
+                    });
+                }
+            },
+            Entry::Vacant(entry) => {
+                entry.insert(AppOrQueue::Queue(
+                    crate::interface::EnqueuedTelemetryData::processed(actions),
+                ));
+            }
+        }
+
+        no_response()
+    }
+
+    type RegisterServiceAndFlushQueuedActionsFut = NoResponse;
+    fn register_service_and_flush_queued_actions(
+        self,
+        _: Context,
+        instance_id: InstanceId,
+        queue_id: QueueId,
+        runtime_meta: RuntimeMetadata,
+        service_name: String,
+        env_name: String,
+    ) -> Self::RegisterServiceAndFlushQueuedActionsFut {
+        // We need a channel to have enqueuing code await
+        let (future, completer) = ManualFuture::new();
+        let app_or_queue = {
+            let rt_info = self.get_runtime(&instance_id);
+            let mut app_or_actions = rt_info.lock_app_or_actions();
+            match app_or_actions.get(&queue_id) {
+                Some(AppOrQueue::Queue(_)) => {
+                    app_or_actions.insert(queue_id, AppOrQueue::App(future.shared()))
+                }
+                None => Some(AppOrQueue::Queue(
+                    crate::interface::EnqueuedTelemetryData::default(),
+                )),
+                _ => None,
+            }
+        };
+        if let Some(AppOrQueue::Queue(mut enqueued_data)) = app_or_queue {
+            tokio::spawn(async move {
+                let mut actions: Vec<TelemetryActions> = vec![];
+                enqueued_data.extract_telemetry_actions(&mut actions).await;
+
+                if let Some(mut app) = self
+                    .get_app(
+                        &instance_id,
+                        &runtime_meta,
+                        &service_name,
+                        &env_name,
+                        actions,
+                    )
+                    .await
+                {
+                    // Register metrics
+                    for metric in std::mem::take(&mut enqueued_data.metrics).into_iter() {
+                        app.register_metric(metric);
+                    }
+
+                    let mut actions: Vec<_> = std::mem::take(&mut enqueued_data.actions);
+
+                    // Send metric points
+                    for point in std::mem::take(&mut enqueued_data.points) {
+                        actions.push(app.to_telemetry_point(point));
+                    }
+
+                    // drop on stop
+                    if actions.iter().any(|action| {
+                        matches!(action, TelemetryActions::Lifecycle(LifecycleAction::Stop))
+                    }) {
+                        self.get_runtime(&instance_id)
+                            .lock_app_or_actions()
+                            .remove(&queue_id);
+                    }
+
+                    app.telemetry.send_msgs(actions).await.ok();
+                    // Ok, we dequeued all messages, now new enqueue_actions calls can handle it
+                    completer.complete((service_name, env_name)).await;
+                }
+            });
+        }
+
+        no_response()
+    }
+
+    type SetSessionConfigFut = Pin<Box<dyn Send + futures::Future<Output = ()>>>;
+
+    fn set_session_config(
+        self,
+        _: Context,
+        session_id: String,
+        config: SessionConfig,
+    ) -> Self::SetSessionConfigFut {
+        let session = self.get_session(&session_id);
+        session.modify_telemetry_config(|cfg| {
+            let endpoint =
+                get_product_endpoint(ddtelemetry::config::PROD_INTAKE_SUBDOMAIN, &config.endpoint);
+            cfg.set_endpoint(endpoint).ok();
+        });
+        session.modify_trace_config(|cfg| {
+            let endpoint = get_product_endpoint(
+                datadog_trace_utils::config_utils::PROD_INTAKE_SUBDOMAIN,
+                &config.endpoint,
+            );
+            cfg.set_endpoint(endpoint).ok();
+        });
+        self.trace_flusher
+            .interval
+            .store(config.flush_interval.as_millis() as u64, Ordering::Relaxed);
+        self.trace_flusher
+            .min_force_flush_size
+            .store(config.force_flush_size as u32, Ordering::Relaxed);
+        self.trace_flusher
+            .min_force_drop_size
+            .store(config.force_drop_size as u32, Ordering::Relaxed);
+
+        session.log_guard.lock().unwrap().replace((
+            log::MULTI_LOG_FILTER.add(config.log_level),
+            log::MULTI_LOG_WRITER.add(config.log_file),
+        ));
+
+        if let Some(completer) = self.self_telemetry_config.lock().unwrap().take() {
+            let config = session
+                .session_config
+                .lock()
+                .unwrap()
+                .as_ref()
+                .unwrap()
+                .clone();
+            tokio::spawn(async move {
+                completer.complete(config).await;
+            });
+        }
+
+        Box::pin(async move {
+            session.shutdown_running_instances().await;
+            no_response().await
+        })
+    }
+
+    type ShutdownRuntimeFut = NoResponse;
+
+    fn shutdown_runtime(self, _: Context, instance_id: InstanceId) -> Self::ShutdownRuntimeFut {
+        let session = self.get_session(&instance_id.session_id);
+        tokio::spawn(async move { session.shutdown_runtime(&instance_id.runtime_id).await });
+
+        no_response()
+    }
+
+    type ShutdownSessionFut = NoResponse;
+
+    fn shutdown_session(self, _: Context, session_id: String) -> Self::ShutdownSessionFut {
+        tokio::spawn(async move { SidecarServer::stop_session(&self, &session_id).await });
+        no_response()
+    }
+
+    type SendTraceV04ShmFut = NoResponse;
+
+    fn send_trace_v04_shm(
+        self,
+        _: Context,
+        instance_id: InstanceId,
+        handle: ShmHandle,
+        headers: SerializedTracerHeaderTags,
+    ) -> Self::SendTraceV04ShmFut {
+        if let Some(endpoint) = self
+            .get_session(&instance_id.session_id)
+            .get_trace_config()
+            .endpoint
+            .clone()
+        {
+            tokio::spawn(async move {
+                match handle.map() {
+                    Ok(mapped) => {
+                        self.send_trace_v04(&headers, mapped.as_slice(), &endpoint);
+                    }
+                    Err(e) => error!("Failed mapping shared trace data memory: {}", e),
+                }
+            });
+        }
+
+        no_response()
+    }
+
+    type SendTraceV04BytesFut = NoResponse;
+
+    fn send_trace_v04_bytes(
+        self,
+        _: Context,
+        instance_id: InstanceId,
+        data: Vec<u8>,
+        headers: SerializedTracerHeaderTags,
+    ) -> Self::SendTraceV04BytesFut {
+        if let Some(endpoint) = self
+            .get_session(&instance_id.session_id)
+            .get_trace_config()
+            .endpoint
+            .clone()
+        {
+            tokio::spawn(async move {
+                self.send_trace_v04(&headers, data.as_slice(), &endpoint);
+            });
+        }
+
+        no_response()
+    }
+
+    type PingFut = Ready<()>;
+
+    fn ping(self, _: Context) -> Self::PingFut {
+        future::ready(())
+    }
+
+    type DumpFut = Pin<Box<dyn Send + futures::Future<Output = String>>>;
+
+    fn dump(self, _: Context) -> Self::DumpFut {
+        Box::pin(crate::dump::dump())
+    }
+
+    type StatsFut = Pin<Box<dyn Send + futures::Future<Output = String>>>;
+
+    fn stats(self, _: Context) -> Self::StatsFut {
+        Box::pin(async move {
+            let stats = self.compute_stats().await;
+            simd_json::serde::to_string(&stats).unwrap()
+        })
+    }
+}
